flask/requirements.txt
flask
pymongo
flask-cors
pandas
windrose
seaborn
matplotlib
python-dotenv